+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Fix filtering for non-recursive `Bucket.ls` being done on the client
+
 ## [1.16.0] - 2022-04-27
 
 This release contains a preview of replication support. It allows for basic